requirements.txt
chardet==5.2.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.59.0
importlib_resources==6.5.2
joblib==1.5.2
kiwisolver==1.4.7
llvmlite==0.43.0
matplotlib==3.9.4
MouseInfo==0.1.3
numba==0.60.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.2.1
psutil==7.1.3
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynndescent==0.5.13
pyparsing==3.2.3
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyRect==0.2.0
PyScreeze==1.0.1
PySide6==6.9.2
PySide6_Addons==6.9.2
PySide6_Essentials==6.9.2
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2025.2
pywin32==311
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
shiboken6==6.9.2
six==1.17.0
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.2
umap-learn==0.5.9.post2
xlsxwriter==3.2.5
zipp==3.23.0
